Find and delete files from harddisk
This python script deletes chosen files from a storage device or another form of drive.

First you need a input file in .xls format, which contains the names of the files that you're searching for. 
Next you need to type in the drive that you want to analyze. For every match on these two, you will be asked if you want to delete the file or not - in that way you'll be able to keep some files and still delete the rest. The script creates a .txt file by the name "Files-to-be-deleted.txt", which contains all the matching files with its paths. 

Sidenote: To make sure that all files from your list are deleted and can't be restored easily again, you can use tools as Windows Cipher. Cipher is a command-line tool that can be use to encrypt and decrypt data on drives that use the NTFS file system and to view the encryption status of files and folders from a Command Prompt.