Carving with RegEx from Excel file
This python script can be used to extract text based on Regular Expression from an Excel file and print it out.

First you need to write the name of the file you want to carve from. Next you can change the name of the output file, which by default is "test.csv".